pyproject.toml
[project]
name = "tradingagent-simplified"
version = "0.1.0"
description = "A simplified implementation of TradingAgents Research paper"
authors = [
    {name = "Group-4"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "yfinance (>=0.2.65,<0.3.0)",
    "langchain-core (>=0.3.68,<0.4.0)",
    "langchain-groq (>=0.3.5,<0.4.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "ta (>=0.11.0,<0.12.0)",
    "pandas-ta (>=0.3.14b,<0.4.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "finta (>=1.3,<2.0)",
    "langgraph (>=0.5.1,<0.6.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-google-genai (>=2.1.7,<3.0.0)",
    "google-genai (>=1.25.0,<2.0.0)",
    "pytest (>=8.4.1,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
